+++ src/app/scripts/compilers/SassCompiler.js
@@ -147,14 +147,14 @@ SassCompiler.prototype.sassCompile = function (file, emitter) {
     var command = self.getSassCmd();
         command += ' ' + argv.join(' ');
 
-    exec(command, {timeout: 5000, maxBuffer: 2000*1024}, function (error, stdout, stderr) {
+    exec(command, {timeout: 5000, maxBuffer: 10000*1024}, function (error, stdout, stderr) {
         if (error !== null) {
             emitter.emit('fail');
             self.throwError(stderr, filePath);
         } else {
             emitter.emit('done');
 
-            //watch import file
+            // watch import file
             common.watchImports('sass', filePath);
 
             // auto add css prefix
@@ -162,7 +162,7 @@ SassCompiler.prototype.sassCompile = function (file, emitter) {
                 common.autoprefix(file);
             }
         }
-            
+
         // do awayls
         emitter.emit('always');
     });
@@ -223,7 +223,7 @@ SassCompiler.prototype.compassCompile = function (file, emitter) {
     // if (settings.sourceMap) {
     //     argv.push('--sourcemap');
     // }
-    
+
     var command = self.getCompassCmd(projectConfig.useSystemCommand) + ' ' + argv.join(' ');
 
     exec(command, {cwd: projectDir, timeout: 5000, maxBuffer: 2000*1024}, function (error, stdout, stderr) {